# Improve performance of Enum's generic IsDefined / GetName / GetValues / GetNames