Eliminates the boxing in IsDefined/GetName/GetValues, in GetValues avoids going through the non-generic Array.SetValue for each value, and in GetNames avoids having to go through RuntimeType's GetEnumNames override.

|    Method |           Toolchain |      Mean | Ratio | Allocated |
|---------- |-------------------- |----------:|------:|----------:|
| IsDefined | \master\corerun.exe |  34.39 ns |  1.00 |      24 B |
| IsDefined |     \pr\corerun.exe |  24.37 ns |  0.70 |         - |
|           |                     |           |       |           |
|   GetName | \master\corerun.exe |  52.31 ns |  1.00 |      24 B |
|   GetName |     \pr\corerun.exe |  25.16 ns |  0.48 |         - |
|           |                     |           |       |           |
|  GetNames | \master\corerun.exe |  37.25 ns |  1.00 |      80 B |
|  GetNames |     \pr\corerun.exe |  22.80 ns |  0.61 |      80 B |
|           |                     |           |       |           |
| GetValues | \master\corerun.exe | 573.06 ns |  1.00 |     224 B |
| GetValues |     \pr\corerun.exe |  19.05 ns |  0.03 |      56 B |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Diagnosers;
using BenchmarkDotNet.Running;
using System;

[MemoryDiagnoser]
public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromAssemblies(new[] { typeof(Program).Assembly }).Run(args);

    private DayOfWeek _value = DayOfWeek.Friday;

    [Benchmark]
    public bool IsDefined() => Enum.IsDefined(_value);

    [Benchmark]
    public string GetName() => Enum.GetName(_value);

    [Benchmark]
    public string[] GetNames() => Enum.GetNames<DayOfWeek>();

    [Benchmark]
    public DayOfWeek[] GetValues() => Enum.GetValues<DayOfWeek>();
}
```