@EgorBo 2020-11-06
@jkotas @stephentoub In case if you are interested: I've just intrinsified `Type.GetTypeCode` for `typeof(x)` or `x.GetType()` args https://github.com/dotnet/runtime/compare/master...EgorBo:opt-type-gettypecode but the jit-diff is not really exciting, just two methods so I guess it's not worth the complexity:
```
(-91.67% of base) : System.Collections.Concurrent.dasm - ConcurrentDictionary`2:IsValueWriteAtomic():bool (7 methods)
(-48.73% of base) : System.Private.CoreLib.dasm - Enum:TryParse(String,bool,bool,byref):bool (6 methods)
```