@stephentoub 2020-11-07
> so I guess it's not worth the complexity

Probably not right now, but it's good to have in our back pocket for the future :)